# src/main/resources/api-docs.yaml
openapi: 3.0.0
info:
  title: Moodi-pie API
  description: 스웨거 API
  version: v1.0.0
servers:
  - url: https://moodipie.shop
  - url: https://운영URL
...
paths:
  /users:
    get:
      tags:
        - 회원
      summary: 마이페이지 조회
      description: 내 정보를 조회합니다.
      operationId: userPage
      responses:
        "200":
          description: "이름 : 김무디, 프로필 사진 : moody"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResUserInfoResponse"
        "400":
          description: 요청 실패
          content:
            application/json:
              examples:
                조회 실패 - userEmail 오류:
                  summary: ""
                  description: userEmail 값이 Null 이라서 플레이리스트 조회에 실패했습니다.
                  value:
                    httpStatus: 400
                    message: NULL 값이 존재합니다.
                    divisionCode: G003
      security:
        - Authorization: []
    put:
      tags:
        - 회원
      summary: 내 정보 수정
      description: 내 정보를 수정합니다.
      operationId: updateUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
        required: true
      responses:
        "200":
          description: 수정 성공
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResUserInfoResponse"
        "400":
          description: 요청 실패
          content:
            application/json:
              examples:
                수정 실패 - 필드 에러:
                  summary: ""
                  description: 필드 조건에 맞지 않아 회원정보 수정에 실패했습니다.
                  value:
                    httpStatus: 400
                    message: 잘못된 형식입니다.
                    divisionCode: G002
      security:
        - Authorization: []
    delete:
      tags:
        - 회원
      summary: 회원 탈퇴
      description: 회원 탈퇴합니다.
      operationId: deleteUser
      responses:
        "200":
          description: 탈퇴 성공
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResUserInfoResponse"
        "400":
          description: 요청 실패
          content:
            application/json:
              examples:
                조회 실패 - userEmail 오류:
                  summary: ""
                  description: userEmail 값이 Null 이라서 플레이리스트 조회에 실패했습니다.
                  value:
                    httpStatus: 400
                    message: NULL 값이 존재합니다.
                    divisionCode: G003
      security:
        - Authorization: []
...
  /token:
    post:
      tags:
        - 회원
      summary: 토큰 갱신
      description: 만료될 토큰을 갱신합니다.
      operationId: refreshToken
      parameters:
        - name: Authorization
          in: header
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        "200":
          description: 토큰 갱신 성공
          content:
            '*/*':
              schema:
                type: string
        "400":
          description: BAD_REQUEST
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiRes"
      security:
        - Authorization: []
...
  /signup:
    post:
      tags:
        - 회원
      summary: 회원가입
      description: 회원에 가입합니다.
      operationId: signup
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
        required: true
      responses:
        "201":
          description: 회원가입 성공
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResCreateUserRequest"
        "400":
          description: 요청 실패
          content:
            application/json:
              examples:
                회원가입 실패 - 필드 에러:
                  summary: ""
                  description: 필드 조건에 맞지 않아 회원가입에 실패했습니다.
                  value:
                    httpStatus: 400
                    message: 잘못된 형식입니다.
                    divisionCode: G002
...
  /logout:
    post:
      tags:
        - 회원
      summary: 로그아웃
      description: 로그아웃합니다.
      operationId: logout
      responses:
        "200":
          description: 로그아웃 성공
          content:
            '*/*':
              schema:
                type: string
        "400":
          description: BAD_REQUEST
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiRes"
      security:
        - Authorization: []
  /login:
    post:
      tags:
        - 회원
      summary: 로그인
      description: 내 정보로 로그인합니다.
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLoginRequest"
        required: true
      responses:
        "200":
          description: 로그인 성공
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserLoginResponse"
        "400":
          description: BAD_REQUEST
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiRes"
...
components:
  schemas:
    ApiRes:
      title: API 응답
      type: object
      properties:
        httpStatus:
          type: integer
          description: 응답코드
          format: int32
        data:
          type: object
          description: 응답 데이터
        message:
          type: string
          description: 응답 메세지
          example: ex.조회에 성공했습니다.
        divisionCode:
          type: string
          description: 오류 구분 코드
          example: 오류시 출력
        errors:
          type: array
          description: 상세 필드 에러
          example: 오류시 출력
          items:
            $ref: "#/components/schemas/FieldErrors"
    FieldErrors:
      type: object
      properties:
        field:
          type: string
        value:
          type: string
        reason:
          type: string
    UpdateUserRequest:
      required:
        - username
      type: object
      properties:
        username:
          type: string
          description: 수정된 이름
          example: 둘리(공백X)
    ApiResUserInfoResponse:
      title: API 응답
      type: object
      properties:
        httpStatus:
          type: integer
          description: 응답코드
          format: int32
        data:
          $ref: "#/components/schemas/UserInfoResponse"
        message:
          type: string
          description: 응답 메세지
          example: ex.조회에 성공했습니다.
        divisionCode:
          type: string
          description: 오류 구분 코드
          example: 오류시 출력
        errors:
          type: array
          description: 상세 필드 에러
          example: 오류시 출력
          items:
            $ref: "#/components/schemas/FieldErrors"
    UserInfoResponse:
      type: object
      properties:
        username:
          type: string
          description: 이름
          example: 홍길동
        email:
          type: string
          description: 이메일
          example: moodipie@gmail.com
...
    CreateUserRequest:
      required:
        - email
        - password
        - username
      type: object
      properties:
        username:
          type: string
          description: 이름
          example: 홍길동
        email:
          type: string
          description: 이메일
          example: moodipie@gmail.com
        password:
          maxLength: 2147483647
          minLength: 8
          type: string
          description: password
          example: teamproject1!
    ApiResCreateUserRequest:
      title: API 응답
      type: object
      properties:
        httpStatus:
          type: integer
          description: 응답코드
          format: int32
        data:
          $ref: "#/components/schemas/CreateUserRequest"
        message:
          type: string
          description: 응답 메세지
          example: ex.조회에 성공했습니다.
        divisionCode:
          type: string
          description: 오류 구분 코드
          example: 오류시 출력
        errors:
          type: array
          description: 상세 필드 에러
          example: 오류시 출력
          items:
            $ref: "#/components/schemas/FieldErrors"
...
    UserLoginRequest:
      required:
        - email
        - password
      type: object
      properties:
        email:
          type: string
          description: 이메일
          example: moodipie@gmail.com
        password:
          type: string
          description: password
          example: teamproject1!
    UserLoginResponse:
      type: object
      properties:
        message:
          type: string
          description: 메세지
          example: 회원가입이 완료되었습니다.
        token:
          type: string
          description: 토큰
          example: asdjasdiaojsdasdjaisd2132i438rdhi2d393
...
  securitySchemes:
    Authorization:
      type: http
      scheme: bearer
      bearerFormat: JWT